Refactor membership routes into own blueprint
