+++ routes/membership/__init__.py
@@ -1,0 +1,4 @@
+from flask import Blueprint
+
+bp = Blueprint("membership", __name__)
+from routes.membership import routes
