+++ templates/base/navbar.html
@@ -118,13 +118,13 @@
         <h4 class="message-header">INVITE</h4>   
         <div class=message-body>{{message.body}}</div>
         <div class="message-button-area">
-          <a href="{{url_for('campaign.accept_invite', 
+          <a href="{{url_for('membership.accept_invite', 
                               campaign_name=message.target_campaign.title, 
                               campaign_id=message.target_campaign.id, 
                               message_id=message.id)}}">
             <button class="message-button">ACCEPT</button>
           </a>
-          <a href="{{url_for('campaign.decline_invite', 
+          <a href="{{url_for('membership.decline_invite', 
                               campaign_name=message.target_campaign.title,
                               campaign_id=message.target_campaign.id, 
                               message_id=message.id)}}">
@@ -171,13 +171,13 @@
         <h4 class="message-header">REQUEST</h4>   
         <div class=message-body>{{message.body}}</div>
         <div class="message-button-area">
-          <a href="{{url_for('campaign.confirm_request', 
+          <a href="{{url_for('membership.confirm_request', 
                               campaign_name=message.target_campaign.title, 
                               campaign_id=message.target_campaign.id, 
                               message_id=message.id)}}">
             <button class="message-button">ACCEPT</button>
           </a>
-          <a href="{{url_for('campaign.deny_request', 
+          <a href="{{url_for('membership.deny_request', 
                               campaign_name=message.target_campaign.title,
                               campaign_id=message.target_campaign.id, 
                               message_id=message.id)}}">
