+++ templates/campaigns.html
@@ -48,7 +48,7 @@
           </button>
         </a>
 
-        <a href="{{ url_for('campaign.join_campaign')}}">
+        <a href="{{ url_for('membership.join_campaign')}}">
           <button class="create-campaign-button">
             + JOIN CAMPAIGN
           </button>
