+++ templates/components/searchbar.html
@@ -4,7 +4,7 @@
   
   Found on the campaigns page, timeline page and timeline edit page
   
-  #}
+#}
 
 <div id="searchbar" class="searchbar-border">
   <div class="searchbar-outer">
